remove jit integration tests
